requirements.txt
Flask==2.2.2
PyYAML==6.0
pyuwsgi
uwsgi
flask-cors
requests
flask_monitoringdashboard
python-dotenv